backend/requirements.txt
asgiref==3.3.4
coverage==5.5
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.1.0
django-autoslug==1.9.8